Fix issue with resource search not being done always (closes #131)
The problem was that sometimes when there were search filters in url, even though componentDidMount called changeFilters action the state didn't actually change (because url filters had the same params as the state), and that's why fetching was not done (in componentWillUpdate).

Now url filters are handled in ENTER_OR_CHANGE_SEARCH_PAGE handler (before componentDidMount) and componentDidMount always does fetching.
