+++ app/state/reducers/searchFiltersReducer.js
@@ -1,6 +1,35 @@
-import isEmpty from 'lodash/isEmpty';
+import { camelizeKeys } from 'humps';
+import pick from 'lodash/pick';
 import moment from 'moment';
 import { handleActions } from 'redux-actions';
+
+import timeUtils from 'utils/timeUtils';
+
+const DATE_FORMAT = 'YYYY-MM-DD';
+const TIME_FORMAT = 'HH:mm';
+
+function getAvailableFilters(availableBetween) {
+  if (!availableBetween) return {};
+  const parts = availableBetween.split(',');
+  if (parts.length !== 2) return {};
+  const start = timeUtils.parseDateTime(parts[0]);
+  const end = timeUtils.parseDateTime(parts[1]);
+  if (!start || !end) return {};
+  return {
+    availableStartDate: start.format(DATE_FORMAT),
+    availableStartTime: start.format(TIME_FORMAT),
+    availableEndDate: end.format(DATE_FORMAT),
+    availableEndTime: end.format(TIME_FORMAT),
+  };
+}
+
+export function parseUrlFilters(queryParams) {
+  const camelized = camelizeKeys(queryParams);
+  const { availableBetween, ...regular } = camelized;
+  const correctKeys = Object.keys(getInitialState());
+  const cleanedRegular = pick(regular, correctKeys);
+  return { ...getAvailableFilters(availableBetween), ...cleanedRegular };
+}
 
 export function getInitialState() {
   const today = moment().format('YYYY-MM-DD');
@@ -25,7 +54,10 @@ export default handleActions({
     ...action.payload,
   }),
   ENTER_OR_CHANGE_SEARCH_PAGE: (state, action) => {
-    if (isEmpty(action.payload.query)) return getInitialState();
-    return state;
+    const urlFilters = parseUrlFilters({ ...action.payload.query });
+    return {
+      ...getInitialState(),
+      ...urlFilters,
+    };
   },
 }, getInitialState());
